Move WorkQueueTaskFailure to workqueue/errors.py
